% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmcFunctions.R
\name{simulateDMC}
\alias{simulateDMC}
\title{Obtain simulated response times and accuracy from the DMC model}
\usage{
simulateDMC(parms, nTrials, dt = 0.1, seed = 42)
}
\arguments{
\item{parms}{The set of parameters to use to simulate the data. Must be
contained in a vector in the order: \code{b}, \code{driftControlled},
\code{maxAmp}, \code{tau}, \code{alpha}, \code{ter}, \code{ster},
\code{sigma}.}

\item{nTrials}{How many trials to simulate per congruency condition.}

\item{dt}{The diffusion scaling parameter (i.e., time steps). By default,
this is set to 0.001.}

\item{seed}{The value for the \code{set.seed} function to set random
generation state.}

\item{var}{The variance of the diffusion process. By default this is set to
0.01.}
}
\value{
Returns a data frame with three columns: rt (response time) in
seconds, accuracy of the model's response (1 for correct, 0 for error), and
congruency condition.
}
\description{
\code{simulateDMC} generates synthetic data from the DMC model in the
form of response time (RT) in seconds and accuracy for both congruent and
incongruent trials.
}
\details{
This function can be employed by the user to generate synthetic data, but
its main purpose is to be used by the fitting procedure to generate model
predictions for a set of parameter values when trying to find the best-
fitting values.d
}
\examples{
# declare the parameters
\dontrun{
parms <- c(50, 0.69, 19.20, 118, 2.15,331,36,3.98)

# simulate the data
modelData <- simulateDMC(parms, nTrials = 10000)
}

}
