% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dstpFunctions.R
\name{fitDSTP}
\alias{fitDSTP}
\title{fit the DSTP model to human data}
\usage{
fitDSTP(data, conditionName, parms = c(0.145, 0.08, 0.1, 0.07, 0.325, 1.3,
  0.24), cdfs = c(0.1, 0.3, 0.5, 0.7, 0.9), cafs = c(0.25, 0.5, 0.75),
  maxParms = c(1, 1, 1, 1, 1, 2, 1), nTrials = 50000)
}
\arguments{
\item{data}{A data frame containing human data. See \code{?exampleData} for
data formatted correctly.}

\item{conditionName}{A string (e.g., "present") which states which condition the
model is to fit currently.}

\item{cdfs}{A vector of quantile values for cumulative distribution functions
to be estimated from the human data. The model will attempt to find the
best-fitting parameters that match this distributional data.}

\item{cafs}{A vector of quantiles for conditional accuracy functions to be
estimated from the human data. The model will attempt to find the best-
fitting parameters that match this distributional data.}

\item{nTrials}{An integer stating how many trials to simulate per iteration
of the fitting cycle for each congruency type.}
}
\description{
\code{fitDSTP} fits the DSTP model to human data.
}
\details{
This function can be employed by the user to find best-fitting parameters of
the DSTP model to human data.
}

